backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
paramiko==3.3.1
pywinrm==0.4.3
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
httpx==0.25.2
pydantic==2.5.0
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
rich==13.7.0
questionary==2.0.1
ipaddress==1.0.23
openpyxl==3.1.2
